This system now works to read data. Need to be able to add protocol file control. And add multithreading support.
