+++ requirements.txt
@@ -1,0 +1,2 @@
+numpy~=1.24.2
+pyserial~=3.5
